feat: add resource type
