+++ udata_front/frontend/__init__.py
@@ -42,7 +42,7 @@ def init_app(app):
     nav.init_app(app)
     theme.init_app(app)
 
-    from . import helpers, error_handlers  # noqa
+    from . import helpers, error_handlers, resource_helpers  # noqa
 
     for view in VIEWS:
         _load_views(app, 'udata_front.views.{}'.format(view))
